separate routes and pages configured
